+++ RightpointLabs.Pourcast.Arduino/TapMonitor.h
@@ -2,6 +2,6 @@ extern "C" {
   void deviceSetup(const char* connectionString);
   void deviceTeardown();
   void deviceProcess(int delay);
-  void deviceSend(int timeSinceLastData, double temperature, double weight, int pulses, const char* kegId);
+  void deviceSend(long timeSinceLastData, double temperature, int weight, long pulses, const char* kegId);
 }
